Refactored ByD SOAP services implementation
