+++ byd_service/soap.py
@@ -17,9 +17,6 @@ sap_auth = SAPAuthentication()
 
 class SOAPServices:
 
-	soap_endpoint = 'https://my350679.sapbydesign.com/sap/bc/srt/scs/sap/manageaccountingentryin'
-	wsdl_path = os.path.join(Path(__file__).resolve().parent.parent, 'manageaccountingentryin.wsdl')
-
 	def __init__(self, ):
 		"""
 			Initialize the SOAP client and authenticate with SAP.
@@ -32,7 +29,4 @@ class SOAPServices:
 		client = Client(self.wsdl_path, transport=transport)
 		client.transport.session.auth = sap_auth.http_authentication()
 
-		# Access the services (operations) provided by the SOAP endpoint
-		soap_client = client.create_service("{http://sap.com/xi/AP/FinancialAccounting/Global}binding", self.soap_endpoint)
-
-		self.client, self.soap_client = client, soap_client
+		self.client = client
