+++ byd_service/wsdl/manageaccountingentryin.wsdl
